+++ NEWS.md
@@ -9,6 +9,8 @@ catalog (ID166).
 * Added functionality for `WriteCatalog` to write genic-nongenic indel
 catalog (ID166) to CSV file.
 
+* Added `cbind` and `[` method for `ID166Catalog`.
+
 ## Updated
 * Updated function `PlotCatalog` to change `options("digits")` temporarily to
 prevent weird y axis labels.